+++ cnext_app/lib/components/WorkingPanel.tsx
@@ -10,11 +10,12 @@ import Pane from "react-split-pane-v2";
 import { RootState } from "../../redux/store";
 import SplitPane from "react-split-pane-v2";
 import { CommandName, ContentType, IMessage, WebAppEndpoint } from "../interfaces/IApp";
-import HotkeyComponent from "./hotkeys/HotKeys";
+import { sendMessage, SocketContext } from "./Socket";
 import TerminalManager from "./terminal-manager/TerminalManager";
 import GitManager from "./git-manager/GitManager";
-import { SocketContext } from "./Socket";
 import OpenAiManager from "./openai-manager/OpenAiManager";
+import HotkeyComponent from "./hotkeys/HotKeys";
+import { Notifier } from "./notifier/Notifier";
 
 const WorkingPanel = () => {
     const socket = useContext(SocketContext);
@@ -37,7 +38,7 @@ const WorkingPanel = () => {
                 type: ContentType.STRING,
                 content: `import mlflow; mlflow.set_tracking_uri("${tracking_uri}")`,
             };
-            socket?.emit(WebAppEndpoint.CodeEditor, JSON.stringify(message));
+            sendMessage(socket, WebAppEndpoint.CodeEditor, message);
         }
     };
 
@@ -90,7 +91,7 @@ const WorkingPanel = () => {
             <TerminalManager />
             <OpenAiManager />
             <HotkeyComponent />
-            {/* <Notifier /> */}
+            <Notifier />
         </StyledWorkingPanel>
     );
 };