+++ cnext_app/lib/components/richoutput-panel/result-panel/libStdInInput.tsx
@@ -1,7 +1,7 @@
 import { Socket } from "socket.io-client";
 import { CommandName, ContentType, IMessage, WebAppEndpoint } from "../../../interfaces/IApp";
 
-const createMessage = (content: string) => {
+export const createMessage = (content: string) => {
     let message: IMessage = {
         webapp_endpoint: WebAppEndpoint.ExecutorManager,
         command_name: CommandName.send_stdin,
@@ -13,8 +13,8 @@ const createMessage = (content: string) => {
     return message;
 };
 
-export const sendMessage = (socket: Socket, content: string) => {
-    const message = createMessage(content);
-    console.log(`${message.webapp_endpoint} send message: `, message);
-    socket?.emit(message.webapp_endpoint, JSON.stringify(message));
-};
+// export const sendMessage = (socket: Socket, content: string) => {
+//     const message = createMessage(content);
+//     console.log(`${message.webapp_endpoint} send message: `, message);
+//     socket?.emit(message.webapp_endpoint, JSON.stringify(message));
+// };